platform/agent-registry/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore==2.14.0
google-cloud-logging==3.8.0
pydantic==2.5.0
requests==2.31.0